refactor: move SwaggerItemSelector to own file
