+++ src/web/src/__tests__/components/WSEditorSwaggerPicker.test.tsx
@@ -1,6 +1,7 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { fireEvent, screen, waitFor } from "@testing-library/react";
-import WSEditorSwaggerPicker, { SwaggerItemSelector } from "../../views/workspace/WSEditorSwaggerPicker";
+import WSEditorSwaggerPicker from "../../views/workspace/WSEditorSwaggerPicker";
+import SwaggerItemSelector from "../../views/workspace/SwaggerItemSelector";
 import { render } from "../test-utils";
 import { workspaceApi, specsApi } from "../../services";
