+++ src/web/src/views/workspace/WSEditorClientConfig.tsx
@@ -23,7 +23,7 @@ import { workspaceApi, specsApi, errorHandlerApi } from "../../services";
 import DoDisturbOnRoundedIcon from "@mui/icons-material/DoDisturbOnRounded";
 import AddCircleRoundedIcon from "@mui/icons-material/AddCircleRounded";
 import { Plane, Resource } from "./WSEditorCommandContent";
-import { SwaggerItemSelector } from "./WSEditorSwaggerPicker";
+import SwaggerItemSelector from "./SwaggerItemSelector";
 import AddRoundedIcon from "@mui/icons-material/AddRounded";
 
 interface WSEditorClientConfigDialogProps {
